requirements.txt
torch
streamlit
Pillow
torchvision
requests
tqdm